+++ language-server/Shared.AnZwDev.ALTools/ALSymbols/ALSyntaxTreeSymbol.cs
@@ -65,9 +65,7 @@ namespace AnZwDev.ALTools.ALSymbols
                     }
                 }
             }
-
         }
-
 
     }
 }