+++ vscode-extension/src/syntaxtreeview/syntaxTreeViewMode.ts
@@ -1,0 +1,4 @@
+export enum SyntaxTreeViewMode {
+    TreeView = 0,
+    ClassView = 1
+}