+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 # Change Log
 
 ## 3.0.52
+ - Issue #440 - [Suggestion] Syntax tokens in the Document Syntax Visualizer 
  - Issue #453 - One Statement Per Line bugfix - line break inside "else begin"
  - Issue #493 - Command to collapse empty brackets into a single line - project processing bugfix
  - Issue #495 - Trim Spaces In Enum Captions/Values
 
 Thank you
+ - jhoek for idea #440
  - NKarolak for finding buig in the issue #453
  - rvanbekkum for finding bug in the issue #493
  - mjmatthiesen for reporting issue #495